Don't rely on truthiness
I ran into this today while needing to shell out to a git command with an empty stdin (but I still needed there to be an open and subsequent close of stdin).

The truthiness check for `options.stdin` fails for empty strings as yabbascript thinks that's equalish to false. This change makes it so that only when the user omits to pass a value for stdin do we skip closing stdin.