Add additional regex pattern for HostDown error code (#308)
* Added new regex pattern to detect HostDown errors

An error in GitHub Desktop was being thrown that was not being appropriately parsed.

In order to properly catch it a new regular expression pattern has been added.

* Added test to check parsing for could not resolve host error

An exra regular expression pattern was added to detect a Git.HostDown error.

In order to make sure it is parsing correctly a test has been added.

* Updated New HostDown regex & test with new line characters

Previously the test and regex were missing the "↵" newline characters where appropriate. They have been added to match the original error's placement of new line characters.

* Updated HostDown regex and test string to use correct escape character newline

Previously a commit was added to properly handle newline characters. The characters used however are not newline escape characters. They have been updated to fix the previous issue.

* Simplify HostDown Regex by ignoring final newline character

Detecting the final newline character is not essential for the match to be made, therefore we can safely remove it.

Co-Authored-By: Daniel-McCarthy <dan.willy.mccarthy@gmail.com>