Add ability to configure architecture to download proper Git binary
The main motivation is to support Dugite module on Windows on ARM64.
Since Git for Windows is not available on this architecture yet,
but Windows 10 allows x86 code to run on the ARM64 version of Windows 10.
We should download 'win32-ia32' binary in this case. Sometimes it would be
nice if the architecture could be configurable.

For example Electron relies on Dugite module. To build Electron
for Windows on ARM64 we use Windows x64 host machine. So during the npm install,
the Dugite script downloads x64 git binary based on host arch. However
that is not compatible with WoA. It needs to force to download x86 binary in this case.